# api/api-spec.yml
openapi: 3.0.3
info:
  title: HR SaaS Backend - Time Off API
  version: 1.0.0
  description: |
    API spec for Time Off feature.
servers:
  - url: http://127.0.0.1:9000
    description: Local
  - url: https://localhost:9000
    description: Local
...
security:
  - bearerAuth: []
paths:
  /api/time-off-requests:
    post:
      tags: [Time Off Requests]
      summary: Create time off request
      description: Employee and company scope are derived from bearer token.
      operationId: createTimeOffRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeOffRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessTimeOffRequestResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [Time Off Requests]
      summary: List time off requests
      operationId: listTimeOffRequests
      parameters:
        - in: query
          name: time_off_type_id
          schema:
            type: string
        - in: query
          name: request_status
          schema:
            $ref: '#/components/schemas/RequestStatus'
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessTimeOffRequestListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RequestStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    TimeOffRequest:
      type: object
      properties:
        id:
          type: string
        employee_id:
          type: string
        time_off_type_id:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        requested_days:
          type: integer
          minimum: 1
        request_reason:
          type: string
        request_status:
          $ref: '#/components/schemas/RequestStatus'
        action_by:
          type: string
          nullable: true
        action_at:
          type: string
          format: date-time
          nullable: true
        action_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateTimeOffRequest:
      type: object
      required:
        - time_off_type_id
        - start_date
        - end_date
        - requested_days
        - request_reason
      properties:
        time_off_type_id:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        requested_days:
          type: integer
          minimum: 1
        request_reason:
          type: string
          maxLength: 255
...
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total_item:
          type: integer
        total_page:
          type: integer
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Unauthorized
    SuccessTimeOffRequestResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TimeOffRequest'
        paging:
          $ref: '#/components/schemas/PageMetadata'
        errors:
          type: string
          nullable: true
    SuccessTimeOffRequestListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TimeOffRequest'
        paging:
          $ref: '#/components/schemas/PageMetadata'
        errors:
          type: string
          nullable: true
...